chore: refactored and added tests to mdtexlib files (#19)
