+++ mdtexlib/src/renderer/mod.rs
@@ -1,5 +1,8 @@
 use crate::ir::BlockToken;
 
+#[cfg(test)]
+mod tests;
+
 pub fn render(ir_list: Vec<BlockToken>) {
     todo!();
 }
